+++ backend/AccountingInventory.Core/DTOs/ProductSpecParams.cs
@@ -2,6 +2,6 @@ namespace AccountingInventory.Core.DTOs
 {
     public class ProductSpecParams : BaseSpecParams
     {
-        public int? ProductId { get; set; }
+        public long? ProductId { get; set; }
     }
 }
