+++ backend/AccountingInventory.Core/DTOs/TransactionDtos.cs
@@ -9,7 +9,7 @@ namespace AccountingInventory.Core.DTOs
     public class CreatePurchaseDetailDto
     {
         [Required]
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         [Required]
         public decimal Quantity { get; set; }
         [Required]
@@ -22,15 +22,15 @@ namespace AccountingInventory.Core.DTOs
         public string PurchaseNo { get; set; } = string.Empty;
         public DateTime Date { get; set; } = DateTime.UtcNow;
         [Required]
-        public int SupplierId { get; set; }
+        public long SupplierId { get; set; }
         public decimal PaidAmount { get; set; }
-        public int PaymentAccountId { get; set; }
+        public long PaymentAccountId { get; set; }
         public List<CreatePurchaseDetailDto> Items { get; set; } = new();
     }
 
     public class PurchaseDetailDto
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public string ProductName { get; set; } = string.Empty;
         public decimal Quantity { get; set; }
         public decimal UnitCost { get; set; }
@@ -39,10 +39,10 @@ namespace AccountingInventory.Core.DTOs
 
     public class PurchaseDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string PurchaseNo { get; set; } = string.Empty;
         public DateTime Date { get; set; }
-        public int SupplierId { get; set; }
+        public long SupplierId { get; set; }
         public string SupplierName { get; set; } = string.Empty;
         public decimal TotalAmount { get; set; }
         public decimal PaidAmount { get; set; }
