+++ backend/AccountingInventory.Infrastructure/Repositories/GenericRepository.cs
@@ -17,7 +17,7 @@ namespace AccountingInventory.Infrastructure.Repositories
             _context = context;
         }
 
-        public async Task<T?> GetByIdAsync(int id)
+        public async Task<T?> GetByIdAsync(long id)
         {
             return await _context.Set<T>().FindAsync(id);
         }
